requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h5py==3.11.0
httplib2==0.22.0
idna==3.7
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
oauth2client==4.1.3
opencv-python-headless==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyDrive==1.3.1
Pygments==2.18.0
pymongo==4.7.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
segment-anything==1.0
six==1.16.0
soupsieve==2.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
